Move trading logic to a service file
